iternal converters, progress text style, disable tooltip for progress
