+++ Notifications.Wpf/Converters/IsNull.cs
@@ -7,7 +7,7 @@ using Notification.Wpf.Converters;
 namespace Notification.Wpf
 {
     [ValueConversion(typeof(object), typeof(bool)), MarkupExtensionReturnType(typeof(IsNull))]
-    public class IsNull : ValueConverter
+    internal class IsNull : ValueConverter
     {
         public override object Convert(object v, Type t, object p, CultureInfo c) => v is null;
