+++ Notifications.Wpf/Converters/NotificationTextTrimmingConverter.cs
@@ -22,14 +22,4 @@ namespace Notification.Wpf
             };
         }
     }
-    [ValueConversion(typeof(NotificationContent), typeof(Visibility)), MarkupExtensionReturnType(typeof(NotificationAttachVisibleConverter))]
-    internal class NotificationAttachVisibleConverter : ValueConverter
-    {
-        public override object Convert(object v, Type t, object p, CultureInfo c)
-        {
-            if (!(v is NotificationContent content)) return Visibility.Collapsed;
-            return content.Message.Length < 43 * content.RowsCount ? Visibility.Collapsed : Visibility.Visible;
-
-        }
-    }
 }
