+++ Notifications.Wpf/Notification.Wpf.csproj
@@ -84,11 +84,14 @@
     <Compile Include="Controls\ProgresBarAnimateBehavior.cs" />
     <Compile Include="Controls\ReversibleStackPanel.cs" />
     <Compile Include="Converters\ActionToCommandConverter.cs" />
+    <Compile Include="Converters\BoolToVisibility.cs" />
     <Compile Include="Converters\Combine.cs" />
     <Compile Include="Converters\IsNull.cs" />
     <Compile Include="Converters\MessageRowsCountConverter.cs" />
+    <Compile Include="Converters\NotificationAttachVisibleConverter.cs" />
     <Compile Include="Converters\NotificationTextAttachConverter.cs" />
     <Compile Include="Converters\NotificationTextTrimmingConverter.cs" />
+    <Compile Include="Converters\StringIsNullOrWhiteSpace.cs" />
     <Compile Include="Converters\ValueConverter.cs" />
     <Compile Include="Extensions\ColumnDefinitionCollapsable.cs" />
     <Compile Include="Extensions\RowDefinitionCollapsable.cs" />
